Create TASK_Number-Guessing-Game
The given code is a Java program for a simple guessing game. It generates a random number between 1 and 100 and prompts the user to guess the number within a limited number of attempts. If the user guesses correctly, a success message is displayed. Otherwise, a failure message is shown along with the correct number. The program keeps track of the total rounds played and total attempts made, and allows the user to play again or exit.